Use quiz attempt id as simulated event's objectid for grader report
The function `create_submission_event_data()` creates simulated event
data to pass to the submission handler, when it is called as part of
the grader report.

For quizzes, it sets the simulated event's objectid to
`$linkarray['area']`, which is the question attempt id (coming from
Moodle core code's `formulation_and_controls()` function).

But the quiz submission handler is meant to receive an instance of
\mod_quiz\event\attempt_submitted, in which the "objectid" field is
the *quiz* attempt ID.